+++ src/world/world.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   world.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: elraira- <elraira-@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: sguilher <sguilher@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/04 14:24:39 by elraira-          #+#    #+#             */
-/*   Updated: 2022/11/27 14:41:220 by elraira-         ###   ########.fr       */
+/*   Updated: 2022/12/03 14:20:02 by sguilher         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
